moved each blueprint in mscolab and mswms into its own directory (including templates)
