+++ mslib/mscolab/blueprints/auth/templates/auth/idp/available_idps.html
