+++ mslib/mscolab/blueprints/auth/templates/auth/idp/idp_login_success.html
