+++ mslib/mscolab/blueprints/auth/templates/auth/user/activate.html
