+++ mslib/mscolab/blueprints/auth/templates/auth/user/confirmed.html
