+++ mslib/mscolab/blueprints/auth/templates/auth/user/reset_confirmation.html
