+++ mslib/mscolab/blueprints/auth/templates/auth/user/reset_password.html
