+++ mslib/mscolab/blueprints/auth/templates/auth/user/status.html
