+++ mslib/mscolab/blueprints/chat/__init__.py
