+++ mslib/mswms/app/__init__.py
@@ -28,14 +28,13 @@ import os
 import logging
 import mslib
 
-from flask import Flask, url_for, render_template, send_from_directory, send_file, abort, Blueprint
+from flask import Flask, url_for
 from xstatic.main import XStatic
-from mslib.msui.icons import icons
+from mslib.mswms.blueprints.docs import DOCS_BP
 
 from mslib.mswms.gallery_builder import STATIC_LOCATION
 from mslib.utils import prefix_route, release_info
 from mslib.utils.file_exists import file_exists
-from mslib.utils.get_content import get_content
 
 DOCS_SERVER_PATH = os.path.dirname(os.path.abspath(mslib.__file__))
 DOCS_STATIC_DIR = os.path.join(DOCS_SERVER_PATH, 'static')
@@ -44,12 +43,6 @@ DOCS_DOCS_DIR = os.path.join(DOCS_STATIC_DIR, 'docs')
 DOCS_TEMPLATES_DIR = os.path.join(DOCS_STATIC_DIR, 'templates')
 # This can be used to set a location by SCRIPT_NAME for testing. e.g. export SCRIPT_NAME=/demo/
 SCRIPT_NAME = os.environ.get('SCRIPT_NAME', '/')
-
-DOCS_BP = Blueprint(
-    "docs",
-    __name__,
-    template_folder=os.path.join(DOCS_TEMPLATES_DIR)
-)
 
 
 class default_mswms_settings:
@@ -138,81 +131,5 @@ def create_app(name="", imprint=None, gdpr=None):
     APP.jinja_env.globals["gdpr"] = gdpr_file
     APP.jinja_env.globals.update(get_topmenu=get_topmenu)
 
-    @DOCS_BP.route('/xstatic/<name>/<path:filename>')
-    def files(name, filename):
-        base_path = _xstatic(name)
-        if base_path is None:
-            abort(404)
-        if not filename:
-            abort(404)
-        return send_from_directory(base_path, filename)
-
-    @DOCS_BP.route('/mss_theme/img/<path:filename>')
-    def mss_theme(filename):
-        base_path = os.path.join(DOCS_IMG_DIR)
-        return send_from_directory(base_path, filename)
-
-    @DOCS_BP.route("/index")
-    def index():
-        return render_template("/index.html")
-
-    @DOCS_BP.route("/mss/about")
-    @DOCS_BP.route("/mss")
-    def about():
-        _file = os.path.join(DOCS_DOCS_DIR, 'about.md')
-        img_url = url_for('docs.overview')
-        md_overrides = ('![image](/mss/overview.png)', f'![image]({img_url})')
-
-        html_overrrides = ('<img alt="image" src="/mss/overview.png" />',
-                           '<img class="mx-auto d-block img-fluid" alt="image" src="/mss/overview.png" />')
-        content = get_content(_file, md_overrides=md_overrides, html_overrides=html_overrrides)
-        return render_template("/content.html", act="about", content=content)
-
-    @DOCS_BP.route("/mss/install")
-    def install():
-        _file = os.path.join(DOCS_DOCS_DIR, 'installation.md')
-        content = get_content(_file)
-        return render_template("/content.html", act="install", content=content)
-
-    @DOCS_BP.route("/mss/help")
-    def help():  # noqa: A001
-        _file = os.path.join(DOCS_DOCS_DIR, 'help.md')
-        html_overrides = ('<img alt="Waypoint Tutorial" '
-                          'src="https://mss.readthedocs.io/en/stable/_images/tutorial_waypoints.gif" />',
-                          '<img  class="mx-auto d-block img-fluid" alt="Waypoint Tutorial" '
-                          'src="https://mss.readthedocs.io/en/stable/_images/tutorial_waypoints.gif" />')
-        content = get_content(_file, html_overrides=html_overrides)
-        return render_template("/content.html", act="help", content=content)
-
-    @DOCS_BP.route("/mss/imprint")
-    def imprint():
-        if file_exists(imprint_file):
-            content = get_content(imprint_file)
-            return render_template("/content.html", act="imprint", content=content)
-        else:
-            return ""
-
-    @DOCS_BP.route("/mss/gdpr")
-    def gdpr():
-        if file_exists(gdpr_file):
-            content = get_content(gdpr_file)
-            return render_template("/content.html", act="gdpr", content=content)
-        else:
-            return ""
-
-    @DOCS_BP.route('/mss/favicon.ico')
-    def favicons():
-        base_path = icons("16x16", "favicon.ico")
-        return send_file(base_path)
-
-    @DOCS_BP.route('/mss/logo.png')
-    def logo():
-        base_path = icons("64x64", "mss-logo.png")
-        return send_file(base_path)
-
-    @DOCS_BP.route('/mss/overview.png')
-    def overview():
-        base_path = os.path.join(DOCS_IMG_DIR, 'wise12_overview.png')
-        return send_file(base_path)
     APP.register_blueprint(DOCS_BP)
     return APP
